+++ packages/stable/src/api/annotations/annotationsApi.ts
@@ -29,6 +29,25 @@ export class AnnotationsAPI extends BaseResourceAPI<AnnotationModel> {
 
   /**
    * [Create annotations](https://docs.cognite.com/api/playground/#operation/annotationsCreate)
+   *
+   * const data = {
+   *   pageNumber: 7,
+   *   textRegion: { xMin: 0, xMax: 0.1, yMin: 0, yMax: 0.2 },
+   *   extractedText: 'i am your father',
+   * };
+   * const partial: AnnotationCreate = {
+   *   annotatedResourceType: 'file',
+   *   annotatedResourceId: 1,
+   *   annotationType: 'documents.ExtractedText',
+   *   creatingApp: 'integration-tests',
+   *   creatingAppVersion: '0.0.1',
+   *   creatingUser: 'integration-tests',
+   *   status: 'suggested',
+   *   data,
+   * };
+   *
+   * const created = await client.annotations.create([partial]);
+   *
    */
   public create = (items: AnnotationCreate[]) => {
     return this.createEndpoint(items);
@@ -36,6 +55,23 @@ export class AnnotationsAPI extends BaseResourceAPI<AnnotationModel> {
 
   /**
    * [Suggest annotations](https://docs.cognite.com/api/playground/#operation/annotationsSuggest)
+   *
+   * const data = {
+   *   pageNumber: 7,
+   *   textRegion: { xMin: 0, xMax: 0.1, yMin: 0, yMax: 0.2 },
+   *   extractedText: 'i am your father',
+   * };
+   * const partial: AnnotationSuggest = {
+   *   annotatedResourceType: 'file',
+   *   annotatedResourceId: 1,
+   *   annotationType: 'documents.ExtractedText',
+   *   creatingApp: 'integration-tests',
+   *   creatingAppVersion: '0.0.1',
+   *   creatingUser: 'integration-tests',
+   *   data,
+   * };
+   *
+   * const created = await client.annotations.suggest([partial]);
    */
   public suggest = (items: AnnotationSuggest[]) => {
     return this.createEndpoint(items, this.suggestUrl);
@@ -45,6 +81,10 @@ export class AnnotationsAPI extends BaseResourceAPI<AnnotationModel> {
    * [Retrieve](https://docs.cognite.com/api/playground/#operation/annotationsByids)
    * or [Get an](https://docs.cognite.com/api/playground/#operation/annotationsGet)
    * annotation
+   *
+   * const annotationIds: InternalId[] = [1,2];
+   *
+   * const response = await client.annotations.retrieve(annotationIds);
    */
   public retrieve = (ids: InternalId[]) => {
     return this.retrieveEndpoint(ids);
@@ -52,6 +92,12 @@ export class AnnotationsAPI extends BaseResourceAPI<AnnotationModel> {
 
   /**
    * [Advanced list of annotations](https://docs.cognite.com/api/playground/#operation/annotationsFilter)
+   *
+   * const limitOne = await client.annotations.list({
+   *   limit: 1,
+   *   filter: fileFilter(annotatedFileId),
+   * });
+   *
    */
   public list = (
     filter: AnnotationFilterRequest
@@ -61,6 +107,10 @@ export class AnnotationsAPI extends BaseResourceAPI<AnnotationModel> {
 
   /**
    * [Delete annotations](https://docs.cognite.com/api/playground/#operation/annotationsDelete)
+   *
+   * const annotationIds: InternalId[] = [1,2];
+   *
+   * await client.annotations.delete(annotationIds);
    */
   public delete = (ids: InternalId[]) => {
     return this.deleteEndpoint(ids);
@@ -68,6 +118,24 @@ export class AnnotationsAPI extends BaseResourceAPI<AnnotationModel> {
 
   /**
    * [Update annotations](https://docs.cognite.com/api/playground/#operation/annotationsUpdate)
+   *
+   * const data = {
+   *   pageNumber: 8,
+   *   fileRef: { externalId: 'def_file_changed' },
+   *   textRegion: { xMin: 0, xMax: 0.1, yMin: 0, yMax: 0.2 },
+   * };
+   * const changes: AnnotationChangeById[] = [
+   *   {
+   *     id: 1,
+   *     update: {
+   *       data: {
+   *         set: data,
+   *       },
+   *     },
+   *   },
+   * ];
+   *
+   * const updatedResp = await client.annotations.update(changes);
    */
   public update = (changes: AnnotationChangeById[]) => {
     return this.updateEndpoint(changes);