+++ packages/beta/CHANGELOG.md
@@ -3,6 +3,94 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.14](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.13...@cognite/sdk-beta@5.2.14) (2022-10-04)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.13](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.12...@cognite/sdk-beta@5.2.13) (2022-09-26)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.12](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.11...@cognite/sdk-beta@5.2.12) (2022-09-23)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.11](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.10...@cognite/sdk-beta@5.2.11) (2022-09-23)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.10](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.9...@cognite/sdk-beta@5.2.10) (2022-09-23)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.9](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.8...@cognite/sdk-beta@5.2.9) (2022-09-12)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.8](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.7...@cognite/sdk-beta@5.2.8) (2022-09-12)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.7](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.6...@cognite/sdk-beta@5.2.7) (2022-09-07)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.6](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.5...@cognite/sdk-beta@5.2.6) (2022-09-07)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.5](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.4...@cognite/sdk-beta@5.2.5) (2022-09-06)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.4](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.3...@cognite/sdk-beta@5.2.4) (2022-08-31)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
 ## [5.2.3](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.2...@cognite/sdk-beta@5.2.3) (2022-08-21)
 
 **Note:** Version bump only for package @cognite/sdk-beta